requirements.txt
aiohttp
requests
python-telegram-bot
beautifulsoup4